refactor: streamline category handling by consolidating default categories and updating TaskCard to use dynamic category colors
